Add Armchair option
Previously, the graphene generated was in the 'zigzag' orientation, which produced odder looking transmission graphs.

So, add the option for the user to have the graphene generated be in the 'armchair' orientation, giving the more well known graphene transmission graph.

Oh, and use lots of ternary operators to do so :)